Add files via upload
For your convenience added a program in folder tools for displaying The Dutch zipcode table. You can use one of this records to add zipcodes in the program. Take the zipcode and a housenumber between lower and upper number to use. The table is very big 9 (ca +300000 records) so it takes a few second to load.